[ 박수로 ] 리팩토링 : 에러핸들링 통합 후 useErrorHandling 훅으로 리팩토링
