+++ src/hooks/useBoard.js
@@ -1,7 +1,6 @@
 import axios from 'axios';
 import useSWR from 'swr';
 import { PUBLIC_IP } from '../config';
-import errorHandling from '../functions/errorHandling';
 export default function useBoard(boardId) {
   const fetcher = (url) =>
     axios.get(url).then((response) => response.data.data);
