+++ src/hooks/useUserInfo.js
@@ -2,7 +2,6 @@ import axios from 'axios';
 import React from 'react';
 import useSWR from 'swr';
 import { PUBLIC_IP } from '../config';
-import errorHandling from '../functions/errorHandling';
 
 export default function useUserInfo(params) {
   const fetcher = (url) =>
