+++ src/views/MyPage/MyPage.js
@@ -9,9 +9,10 @@ import { withRouter } from 'react-router';
 import Page404 from '../Page404/Page404';
 import Header1 from '../Common/Header';
 import styles from '../../css/MyPage.module.css';
-import errorHandling from '../../functions/errorHandling';
 import useUserInfo from '../../hooks/useUserInfo';
+import useErrorHandling from '../../hooks/useErrorHandling';
 function MyPage(props) {
+  const errorHandling = useErrorHandling();
   const { Header, Content, Footer } = Layout;
   const { user, isError, isLoading } = useUserInfo();
